refactor: moved session capabilities to session types
